feat: finish scheduled job part of task, move config to controller annotations

+++ src/main/java/betr/intern/chainsaw/service/UserService.java
@@ -50,7 +50,11 @@ public class UserService {
     }
 
     @Transactional
-    public void deleteById(final UUID id) {
+    public String deleteById(final UUID id) {
+        if (userRepository.existsById(id)) {
+            return String.format("User with id=%s did not exist to begin with", id);
+        }
         userRepository.deleteById(id);
+        return String.format("User with id=%s deleted", id);
     }
 }
